+++ NEWS.md
@@ -4,7 +4,8 @@
 
 * `get_sprint()` now returns a tibble instead of a data.frame, consistent 
 with other functions
-* `get_issue()` returns the issue summary, description, and resolution
+* `get_issue()` returns also returns the issue summary, description, and resolution
+* `get_issue()` checks if the story point mapping has been refreshed in the current session and will auto refresh if necessary. This behavior can be disabled by setting the option "sprintr_storypoint_mapping" to TRUE.
 
 
 # sprintr 0.1.2